Run a targeted terraform plan
A full plan takes several minutes, which is a long
time to keep users waiting for feedback on their
PRs. This change uses a shell script to run a 
targeted terraform plan against each module (i.e.
repository) which is changed by the current PR.

NB: This will break if the terraform/*.tf filename
does not match the module name of the repository
collaborators.