WiP: github action to compare github & terraform
At this stage, I just want to check that I can get
the terraform state via a github action.

I'm trying to do this by invoking the AWS CLI via
`docker run`, so that I don't have to build a
docker image for this project with the AWS SDK
ruby gems in it. That's my fallback if I can't get
it to work this way.