Add a class to return data from the terraform src
This is a bit icky and possibly fragile, because
it's using regular expressions to pull data out of
the terraform source code, which means it's
dependent on the format staying stable.

But a) there doesn't seem to be any 'standard' way
for ruby to parse terraform source code, and b)
we are running `terraform fmt` via the code-
formatter github action, so the code should be
consistently formatted.

So this will do as a first iteration, and we can
replace it with something smarter later on.

NB: The data we need is not present in the
terraform state, only in the source, which is why
we're doing it this way.