Refactor graphql code - re-use pagination
This adds a `get_paginated_results` method to the
GraphQlClient class, and calls that to list org.
members and repositories.

This DRYs up the graphql code, but it means the
specs become much harder to write.

For now, I've removed the relevant specs. This is
not ideal, but I think this change is worth 
making anyway.