Allow comments in terraform source
Prior to this change, there was an implicit
requirement for any supplied values to have
nothing after the closing quote. So, this would be
fine:

```
   foo = "bar"
```

...but this would count as missing the value of
`foo`:

```
   foo = "bar" # some comment here
```

With this change, such comments will be ignored,
as they should be.

Although other text on the same line will also be
allowed, the combination of `terraform fmt` and
`terraform plan` which runs on every PR should 
prevent malformed code from getting through.