Use DynamoDB table to lock terraform state
Depends on https://github.com/ministryofjustice/cloud-platform-environments/pull/4029

Once that PR is merged, the AWS credentials stored
in the github secrets `AWS_ACCESS_KEY_ID` and
`AWS_SECRET_ACCESS_KEY` must be replaced with the
credentials belonging to the IAM user created by
that PR.

The new IAM user has access to both the S3 bucket
and the DynamoDB table, which is required to use
terraform state locking.

The previous AWS credentials only had access to
the S3 bucket.