Move the terraform state to a new path in S3
This reflects a change I made manually.

Having the terraform.tfstate at the top-level of
the S3 bucket was a bad idea. This isn't the only
terraform state file we're ever going to have in
that S3 bucket, so all terraform.tfstate files
need a key prefix in the bucket, so that we know
what they are, and don't end up overwriting them
by accident.