Remove undefined collabs after report is updated
This is the most sensible time to do this, since
we know at this point that the report is up to
date.

This does mean that the report will show out of
date information all day - listing collaborators
who have already been removed.

This is probably a good thing, because it provides
an easy way to check, if someone complains that
they have lost access to a repo. If they're listed
in the report, they were probably removed by the
script.