Remove TerraformCollaborator#remove method
This method was originally defined on this class,
but I removed it because it doesn't make sense for
it to be there. We parse the JSON report to find 
undefined collaborators, so this method doesn't
belong on this class.

I removed this code before, but I must have 
accidentally restored it, probably during a merge.