Replace `filter` with `find_all`
The `filter` method is a synonym for `find_all`, but it's only present
in ruby 2.7+ (I think). This script runs as part of a github action, and
apparently that's using an older version of ruby, because we're getting
these errors:

```
bin/remove-undefined-collaborators.rb:34:in `<main>': undefined method
`filter' for []:Array (NoMethodError)
Error: Process completed with exit code 1.
```

This tweak should fix that.