Remove terraform code defining 3 collaborations
For some reason, these collaborations are being repeatedly destroyed and
recreated, as per #106

This change removes the terraform source, which should completely remove
the collaborations from the terraform state. After this, a later PR will
reinstate them in the hopes that they "stick" this time.

> I have a theory that this can happen if collaboration invitations are
"pending" throughout multiple terraform apply steps. If this works, I
might try to recreate this problem in a different repo, to confirm.